Added a bunch of sorting functions and a new benchmarking sln to test those sorting functions (#19)
* Added selection sort, insertion sort(linear), hoare's quicksort, and basic merge sort(poorly implemented so it's very very slow).

* Added mergesort and quicksort.

* Added custom sort function

* Made some basic math functions branchless(I think) and changed the base case for quick sort and merge sort to be insertion sort.

* Added branchless quicksort and updated benchmarking for sorting.

* Add comparison functors to parameters of the sorting algorithms

* Added optimistic insertion sorting and rectified some compiler warnings

* Add a new sln just for benchmarking and moved all the sorting benchmarks over to there.

+++ Tools/RuhrValley/Profiler.cpp
@@ -117,7 +117,7 @@ void BenchMarker::LogResults()
 		}
 	}
 
-	long double mean = total / m_runTimes.size();
+	long double mean = (long double)total / (long double)m_runTimes.size();
 
 	long double deviation = 0;
 	for(size_t i = 0; i < m_runTimes.size(); i++)
